feat: buffer writer responses to catch marshalling errors
Previously, the response status code was set before the response was assembled. If the writer threw an error, this status code would erroneously prevail. Now, a marshalling errors panics before setting the status code, allowing the panic receiver to gracefully handle the situation with a correct status code (e.g. 500)